Rename settings.from_file to settings.load_localsettings
